fix(codemode): preserve column-capped eval output
Keep raw eval bytes recoverable when the per-line cap fires before the spill threshold.

Teach eval callers to use bounded or offset-based output for large text.

Ultraworked with [omo](https://github.com/code-yeongyu/oh-my-openagent)
